added comments and refact code
